chore: split compositor benchmarks to respect the 500-line cap
The plan-008 keeper benchmark (#589) pushed benchmark_test.go to 530
lines, breaking CI's file-length guard on main (masked at review time
by the nested-worktree lint artifact that also suppresses
check-file-length). Move the style + cache-hit/miss benchmarks into
benchmark_style_test.go; benchmark_test.go returns to 340 lines. No
benchmark content changed.
